Refine Montgomery inner product preparation
